More bringing up-to-speed with the master branch
- Simplify linting with a gulp streaming linter
- Fix order of some subscribable includes
- Add `.editorconfig`
- Break out and simplify `gulpfiles.js` into `gulp/`

Note: Release, and Closure build are not yet tested; Hoping to move watch+testing to Karma shortly
