+++ package.json
@@ -32,9 +32,11 @@
     "gulp-exec": "^1.0.4",
     "gulp-footer": "^1.0.5",
     "gulp-header": "^1.2.2",
+    "gulp-help": "^1.3.4",
     "gulp-inject": "^0.4.1",
     "gulp-jasmine": "brianmhunt/gulp-jasmine",
     "gulp-jshint": "^1.9.0",
+    "gulp-lintspaces": "^0.2.3",
     "gulp-livereload": "^2.1.1",
     "gulp-load-plugins": "^0.7.1",
     "gulp-notify": "^2.0.1",
